Moving tests into ExistingSite namespace.
