+++ docroot/modules/custom/prisoner_hub_entity_access/tests/src/ExistingSite/PrisonerHubQueryAccessTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\prisoner_hub_entity_access\Functional;
+namespace Drupal\Tests\prisoner_hub_entity_access\ExistingSite;
 
 use Drupal\Component\Serialization\Json;
 use Drupal\Core\Url;
